+++ Application/Features/Books/Commands/DeleteBook/DeleteBookHandler.cs
@@ -10,6 +10,10 @@ public class DeleteBookHandler(IBookRepository repository)
     public async Task<Result> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
     {
         var book = await repository.GetByIdAsync(request.Id, cancellationToken);
+        if (book == null)
+        {
+            return Result.Failed("Book not found", ErrorTypeCode.NotFound);
+        }
         repository.Delete(book);
          return Result.Successful();
     }
